LayoutTests/platform/chromium/fast/forms/calendar-picker/week-picker-key-operations.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../../../fast/js/resources/js-test-pre.js"></script>
<script src="../../../../../fast/forms/resources/picker-common.js"></script>
<script src="resources/calendar-picker-common.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>
<input type=week id=week value="2000-W01">
<script>
description("Tests if month picker key bindings work as expected.");

debug('Check that page popup doesn\'t exist at first.');
shouldBeNull('document.getElementById("mock-page-popup")');

openPicker(document.getElementById('week'), test1);

function test1() {
    debug('Check that page popup exists.');
    shouldBe('popupWindow.pagePopupController.toString()', '"[object PagePopupController]"');

    shouldBeEqualToString('selectedWeek()', '2000-W01');
    shouldBeEqualToString('currentMonth()', '2000-01');

    debug('Check that arrow keys work properly.');

    eventSender.keyDown('upArrow');
    shouldBeEqualToString('selectedWeek()', '1999-W52');
    shouldBeEqualToString('currentMonth()', '2000-01');
    
    eventSender.keyDown('upArrow');
    shouldBeEqualToString('selectedWeek()', '1999-W51');
    shouldBeEqualToString('currentMonth()', '1999-12');

    eventSender.keyDown('downArrow');
    shouldBeEqualToString('selectedWeek()', '1999-W52');
    shouldBeEqualToString('currentMonth()', '1999-12');

    eventSender.keyDown('downArrow');
    shouldBeEqualToString('selectedWeek()', '2000-W01');
    shouldBeEqualToString('currentMonth()', '2000-01');
    
    eventSender.keyDown('downArrow');
    shouldBeEqualToString('selectedWeek()', '2000-W02');
    shouldBeEqualToString('currentMonth()', '2000-01');

    eventSender.keyDown('leftArrow');
    shouldBeEqualToString('selectedWeek()', '2000-W01');
    shouldBeEqualToString('currentMonth()', '2000-01');

    eventSender.keyDown('rightArrow');
    shouldBeEqualToString('selectedWeek()', '2000-W02');
    shouldBeEqualToString('currentMonth()', '2000-01');

    debug("press m");
    eventSender.keyDown('m');
    shouldBeEqualToString('selectedWeek()', '2000-W07');
    shouldBeEqualToString('currentMonth()', '2000-02');

    debug("press shift + m");
    eventSender.keyDown('M');
    shouldBeEqualToString('selectedWeek()', '2000-W02');
    shouldBeEqualToString('currentMonth()', '2000-01');

    debug("press y");
    eventSender.keyDown('y');
    shouldBeEqualToString('selectedWeek()', '2001-W03');
    shouldBeEqualToString('currentMonth()', '2001-01');

    debug("press shift + y");
    eventSender.keyDown('Y');
    shouldBeEqualToString('selectedWeek()', '2000-W02');
    shouldBeEqualToString('currentMonth()', '2000-01');

    debug("press d");
    eventSender.keyDown('d');
    shouldBeEqualToString('selectedWeek()', '2010-W02');
    shouldBeEqualToString('currentMonth()', '2010-01');

    debug("press shift + d");
    eventSender.keyDown('D');
    shouldBeEqualToString('selectedWeek()', '2000-W02');
    shouldBeEqualToString('currentMonth()', '2000-01');

    eventSender.keyDown('\n');
    waitUntilClosing(test1AfterClosing);
}

function test1AfterClosing() {
    shouldBeNull('document.getElementById("mock-page-popup")');
    shouldBe('document.getElementById("week").value', '"2000-W02"');

    openPicker(document.getElementById('week'), test2);
}

function test2() {
    // Test twice in case midnight.
    shouldBeTrue('testToday() || testToday()');

    debug('Check if tabbing works.');
    shouldBe('focusedElement()', '".days-area-container"');
    eventSender.keyDown('\t');
    shouldBe('focusedElement()', '".today-button[value=<<ThisWeekLabel>>]"');
    eventSender.keyDown('\t');
    shouldBe('focusedElement()', '".clear-button[value=<<CalendarClear>>]"');
    eventSender.keyDown('\t');
    shouldBe('focusedElement()', '".year-month-button"');
    eventSender.keyDown('\t', ['shiftKey']);
    shouldBe('focusedElement()', '".clear-button[value=<<CalendarClear>>]"');

    eventSender.keyDown('\x1B');
    waitUntilClosing(test2AfterClosing);
}

function test2AfterClosing() {
    shouldBeNull('document.getElementById("mock-page-popup")');
    shouldBe('document.getElementById("week").value', '"2000-W02"');

    sendKey(document.getElementById('week'), "F4");
    if (document.getElementById('mock-page-popup'))
        testPassed("F4 opened picker.");

    finishJSTest();
}

function testToday() {
    eventSender.keyDown('t');
    var now = new Date();
    var expectedWeek = popupWindow.Week.createFromToday().toString();
    return selectedWeek() === expectedWeek;
}

function focusedElement() {
    var element = popupWindow.document.activeElement;
    if (!element)
        return null;
    var identifier = "." + element.className;
    if (element.value)
        identifier += "[value=" + element.value + "]";
    return identifier;
}

</script>
<script src="../../../../../fast/js/resources/js-test-post.js"></script>
</body>
</html>
